Add deposit test for Account and pass test
added test to ensure that after a customer deposits money, their balance increases by the amount deposited into the account. To pass this test, an Account class was created with a balance state, a getter method (getBalance) and a deposit method. the balance of the account on instantiation is set to 0 and the deposit method accepts an argument(amount) and increases the balance by the amount passed in.