Add test for printstatment and transaction date and pass
added test to accountSpec to ensure the dat of transaction shows when the printStatement function is invoked. To pass this test I define a printStatement method in the Account class and added implementation code that called a method of the transaction instance. This triggered the need for a Transaction unit test, so I added a unit test to transactionsSpec.js to ensure that the correct date is returned when the getDate method is called. This led to the passing of both tests. However, the date format was not what i required, to fix this i created a private method in the Transaction class (outputDate) and used this method inside the getDate method.